Trying to fix bangscores #3.6

+++ src/main/java/DiscordBot/helpers/GetBangScores.java
@@ -37,7 +37,7 @@ public class GetBangScores {
             mostAttempts = getMostAttempts.executeQuery();
             mostAttempts.last();
             System.out.println("tries: "+mostAttempts.getDouble("tries"));
-            System.out.println("user: "+mostAttempts.getDouble("user"));
+            System.out.println("user: "+mostAttempts.getFloat("user"));
             attemptCount = mostAttempts.getDouble("tries");
             mostAttemptsPlayer = guild.getMemberById(mostAttempts.getLong("user")).getUser().getName();
  System.out.println("Here2");
@@ -46,7 +46,7 @@ public class GetBangScores {
             mostDeaths = getMostDeaths.executeQuery();
             mostDeaths.next();
             deathCount = mostDeaths.getDouble("deaths");
-            mostDeathsPlayer = guild.getMemberById((long)mostDeaths.getDouble("user")).getUser().getName();
+            mostDeathsPlayer = guild.getMemberById(mostDeaths.getLong("user")).getUser().getName();
  System.out.println("here3");
             // Get luckiest and unluckiest players
             PreparedStatement getLuckRanks = conn.prepareStatement("SELECT user, death_rate FROM bang WHERE "+date.getTime()+" - last_played < 604800000 GROUP BY user, death_rate ORDER BY death_rate");
@@ -55,12 +55,12 @@ public class GetBangScores {
             // Luckiest
             luck.first();
             bestRate = 100 - (Math.round(luck.getDouble("death_rate") * 10d) / 10d);
-            luckiestPlayer = guild.getMemberById((long)luck.getDouble("user")).getUser().getName();
+            luckiestPlayer = guild.getMemberById(luck.getLong("user")).getUser().getName();
 
             // Unluckiest
             luck.last();
             worstRate = 100 - (Math.round(luck.getDouble("death_rate") * 10d) / 10d);
-            unluckiestPlayer = guild.getMemberById((long)luck.getDouble("user")).getUser().getName();
+            unluckiestPlayer = guild.getMemberById(luck.getLong("user")).getUser().getName();
         }
         catch (Exception e) {
             System.out.println("GetBangScores Exception 2");
